+++ src/evidence_collection/universe/__init__.py
@@ -1,4 +1,7 @@
 from .sp500 import fetch_sec_companies, fetch_sp500, fetch_sp500_with_ciks
+from .domains import apply_domains, domain_for_ticker, load_company_domains
+from .aliases import load_company_aliases, seed_company_aliases
+from .load import enrich_companies, load_universe
 from .entity import (
     QUERY_ALIASES,
     clean_company_name,
@@ -11,6 +14,13 @@ __all__ = [
     "fetch_sp500",
     "fetch_sp500_with_ciks",
     "fetch_sec_companies",
+    "apply_domains",
+    "domain_for_ticker",
+    "load_company_domains",
+    "load_company_aliases",
+    "seed_company_aliases",
+    "load_universe",
+    "enrich_companies",
     "QUERY_ALIASES",
     "clean_company_name",
     "search_name",